refactor(tests): reuse common helper functions
